Task1/src/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
filelock==3.16.1
gdown==5.2.0
idna==3.10
isodate==0.7.2
msrest==0.7.1
numpy==2.0.2
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pip==24.2
PySocks==1.7.1
pytesseract==0.3.13
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
uvicorn==0.16.0
gunicorn
fastapi==0.70.1
python-multipart
matplotlib
symspellpy